requirements.txt
absl-py==1.4.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.0
astunparse==1.6.3
attrs==22.1.0
autopep8==1.5.7
awsfed==1.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.3.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
chess==1.3.0
click==8.1.3
colorama==0.4.4
configparser==5.0.0
contourpy==1.0.6
crayons==0.4.0
cryptography==3.0
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.6.0
distro==1.5.0
dnspython==2.0.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
ghp-import==2.1.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==2.10
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.12.0
importlib-resources==5.10.0
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.3
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
keras==2.11.0
keyring==23.0.1
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mergedeep==1.3.4
mistune==2.0.4
mkdocs==1.3.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==24.0.1
requests==2.24.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.6.0
typing_extensions==4.5.0
urllib3==1.25.10
vzmi.awsfed==1.4.24263684
watchdog==2.1.9
wcwidth==0.2.5
webdriver-manager==3.3.0
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.4.1
